+++ zacboot-admin/src/main/java/com/zacboot/admin/service/impl/SysPermissionServiceImpl.java
@@ -71,7 +71,7 @@ public class SysPermissionServiceImpl extends SysBaseServiceImpl<SysPermissionMa
     public PageResult<SysPermission> queryPage(PermissionRequest permissionRequest) {
         PageResult<SysPermission> pageResult = new PageResult<>();
         pageResult.setDataList(sysPermissionDao.queryPage(permissionRequest,new Page(permissionRequest.getPageNumber(),permissionRequest.getPageSize())));
-        pageResult.setTotal(sysPermissionDao.queryPageCount(permissionRequest).intValue());
+        pageResult.setTotal(sysPermissionDao.queryPageCount(permissionRequest));
         return pageResult;
     }
